Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    float a,b;
+    int a,b;
     char op;
-    scanf("%f %f %c",&a,&b,&op);
+    scanf("%d %d %c",&a,&b,&op);
     switch (op) {
         case '+':
             printf("%d",a+b);
